refactor: Extract <.place /> to a separate file
